VS code insists that my script isn't updated.
VS code insists that my script isn't updated, and keeps using an old version of my code because I have opened/closed GIThub desktop while working.

THIS HAS TO RELATE TO THE ISSUE I had regarding not being able to get into my github account when I moved onto my new laptop. GITHUB is insisting there is a newer version of the code locally, and indicates the changes, but isn't updating the scripts and page to reflect the changes. (not enough changes?)